# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API para gestión de solicitudes de crédito",
        "title": "Bravo Credit Challenge API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.linkedIn.com/in/sirpyerre",
            "email": "pedro.rojas@gmail.com"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/api/v1/applications": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "applications"
                ],
                "summary": "Listar solicitudes",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filtro por país (MX, BR, ES…)",
                        "name": "country",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filtro por estado (PENDING, APPROVED, DENIED…)",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Fecha inicio YYYY-MM-DD",
                        "name": "from_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Fecha fin YYYY-MM-DD",
                        "name": "to_date",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Límite (default 20, max 100)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_Sirpyerre_bravo-challenge_internal_service.ListApplicationsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "applications"
                ],
                "summary": "Crear solicitud de crédito",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Clave de idempotencia (UUID)",
                        "name": "Idempotency-Key",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Datos de la solicitud",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_Sirpyerre_bravo-challenge_internal_service.CreateApplicationRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_Sirpyerre_bravo-challenge_internal_domain.Application"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/applications/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "applications"
                ],
                "summary": "Obtener solicitud por ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID de la solicitud",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_Sirpyerre_bravo-challenge_internal_domain.Application"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "applications"
                ],
                "summary": "Actualizar estado de solicitud",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Clave de idempotencia (UUID)",
                        "name": "Idempotency-Key",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "UUID de la solicitud",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Nuevo estado",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_Sirpyerre_bravo-challenge_internal_service.UpdateApplicationRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/auth/login": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Login",
                "parameters": [
                    {
                        "description": "Credenciales",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_Sirpyerre_bravo-challenge_internal_service.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_Sirpyerre_bravo-challenge_internal_service.AuthResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/auth/register": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Registrar usuario",
                "parameters": [
                    {
                        "description": "Datos de registro",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_Sirpyerre_bravo-challenge_internal_service.RegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_Sirpyerre_bravo-challenge_internal_service.AuthResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "github_com_Sirpyerre_bravo-challenge_internal_domain.Application": {
            "type": "object",
            "properties": {
                "country": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "full_name": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "identity_document": {
                    "type": "string"
                },
                "monthly_income": {
                    "type": "number"
                },
                "notes": {
                    "type": "string"
                },
                "requested_amount": {
                    "type": "number"
                },
                "risk_level": {
                    "$ref": "#/definitions/github_com_Sirpyerre_bravo-challenge_internal_domain.RiskLevel"
                },
                "status": {
                    "$ref": "#/definitions/github_com_Sirpyerre_bravo-challenge_internal_domain.ApplicationStatus"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "github_com_Sirpyerre_bravo-challenge_internal_domain.ApplicationStatus": {
            "type": "string",
            "enum": [
                "PENDING",
                "VALIDATING",
                "APPROVED",
                "DENIED"
            ],
            "x-enum-varnames": [
                "StatusPending",
                "StatusValidating",
                "StatusApproved",
                "StatusDenied"
            ]
        },
        "github_com_Sirpyerre_bravo-challenge_internal_domain.RiskLevel": {
            "type": "string",
            "enum": [
                "LOW",
                "MEDIUM",
                "HIGH"
            ],
            "x-enum-varnames": [
                "RiskLow",
                "RiskMedium",
                "RiskHigh"
            ]
        },
        "github_com_Sirpyerre_bravo-challenge_internal_service.AuthResponse": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                },
                "user_id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                }
            }
        },
        "github_com_Sirpyerre_bravo-challenge_internal_service.CreateApplicationRequest": {
            "type": "object",
            "properties": {
                "country": {
                    "type": "string",
                    "example": "MX"
                },
                "full_name": {
                    "type": "string",
                    "example": "Juan Pérez"
                },
                "identity_document": {
                    "type": "string",
                    "example": "CURP123456"
                },
                "monthly_income": {
                    "type": "number",
                    "example": 15000
                },
                "requested_amount": {
                    "type": "number",
                    "example": 50000
                }
            }
        },
        "github_com_Sirpyerre_bravo-challenge_internal_service.ListApplicationsResponse": {
            "type": "object",
            "properties": {
                "applications": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_Sirpyerre_bravo-challenge_internal_domain.Application"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "github_com_Sirpyerre_bravo-challenge_internal_service.LoginRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "user@example.com"
                },
                "password": {
                    "type": "string",
                    "example": "secret123"
                }
            }
        },
        "github_com_Sirpyerre_bravo-challenge_internal_service.RegisterRequest": {
            "type": "object",
            "properties": {
                "country": {
                    "type": "string",
                    "example": "MX"
                },
                "email": {
                    "type": "string",
                    "example": "user@example.com"
                },
                "password": {
                    "type": "string",
                    "example": "secret123"
                },
                "role": {
                    "type": "string",
                    "example": "USER"
                }
            }
        },
        "github_com_Sirpyerre_bravo-challenge_internal_service.UpdateApplicationRequest": {
            "type": "object",
            "properties": {
                "notes": {
                    "type": "string",
                    "example": "Aprobado por evaluación de riesgo"
                },
                "status": {
                    "type": "string",
                    "example": "APPROVED"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}